update example to work with more recent versions of forcelayout and physics.simulator
When using the original code with the more recent versions of forcelayout and physics.simulator I found that 1) there was a javascript error being thrown from interaction with ngraph.event and 2) the physics parameters I had specified were being ignored. The proposed changes seem to fix both of these issues.

As a new user of ngraph I naturally look to the example code for my own starting point, thus I think it would be generally helpful to the community of users to update the example code to work with the most recent versions of other ngraph components.

Thanks!  ~ry